# swagger.yaml
openapi: 3.0.0
info:
    title: 'API Documentation'
    version: 1.0.0
    description: 'API documentation for plans, workouts, and exercises.'
servers:
    -
        url: 'http://localhost:3010'
paths:
...
    '/plans/{plan_id}/workouts':
        get:
            summary: 'Retrieve all workouts for a specific plan'
            tags: [Plans]
            parameters: [{name: plan_id, in: path, required: true, description: 'ID of the training plan', schema: {type: integer}}]
            responses: {'200': {description: 'A list of workouts for the specified plan', content: {application/json: {schema: {type: array, items: {type: object, properties: {id: {type: integer, example: 1}, plan_id: {type: integer, example: 1}, name: {type: string, example: 'Morning Cardio'}, length: {type: integer, example: 30}, type: {type: string, example: Cardio}, frequency: {type: integer, example: 3}}}}}}}, '403': {description: 'Access denied. Only users and admins can access this endpoint.'}, '404': {description: 'Plan not found'}}
...
    /workouts:
        get:
            summary: 'Retrieve a list of all workouts'
            tags: [Workouts]
            security: [{BearerAuth: []}]
            responses: {'200': {description: 'A list of workouts', content: {application/json: {schema: {type: array, items: {type: object, properties: {id: {type: integer, example: 1}, plan_id: {type: integer, example: 1}, name: {type: string, example: 'Morning Cardio'}, length: {type: integer, example: 30}, type: {type: string, example: Cardio}, frequency: {type: integer, example: 3}}}}}}}, '401': {description: 'Access denied. No token provided.'}, '403': {description: 'Access denied. You do not have the required permissions.'}}
        post:
            summary: 'Create a new workout for a specific training plan'
            tags: [Workouts]
            requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {name: {type: string, example: 'Morning Cardio'}, length: {type: integer, example: 30}, type: {type: string, example: Cardio}, frequency: {type: integer, example: 3}, plan_id: {type: integer, example: 1}}}}}}
            responses: {'201': {description: 'Workout created successfully', content: {application/json: {schema: {type: object, properties: {id: {type: integer, example: 1}, plan_id: {type: integer, example: 1}, name: {type: string, example: 'Morning Cardio'}, length: {type: integer, example: 30}, type: {type: string, example: Cardio}, frequency: {type: integer, example: 3}}}}}}, '400': {description: 'All fields are required'}, '403': {description: 'Access denied. You are not authorized to create a workout for this plan.'}, '422': {description: 'Invalid input values'}}
    '/workouts/{workout_id}':
        get:
            summary: 'Retrieve a specific workout by ID'
            tags: [Workouts]
            parameters: [{name: workout_id, in: path, required: true, description: 'ID of the workout', schema: {type: integer}}]
            responses: {'200': {description: 'A workout found', content: {application/json: {schema: {type: object, properties: {id: {type: integer, example: 1}, plan_id: {type: integer, example: 1}, name: {type: string, example: 'Morning Cardio'}, length: {type: integer, example: 30}, type: {type: string, example: Cardio}, frequency: {type: integer, example: 3}}}}}}, '403': {description: 'Access denied. You are not authorized to view this workout.'}, '404': {description: 'Workout not found'}}
        put:
            summary: 'Update a specific workout'
            tags: [Workouts]
            parameters: [{name: workout_id, in: path, required: true, description: 'ID of the workout', schema: {type: integer}}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {name: {type: string, example: 'Morning Cardio'}, length: {type: integer, example: 30}, type: {type: string, example: Cardio}, frequency: {type: integer, example: 3}}}}}}
            responses: {'200': {description: 'Workout updated successfully'}, '400': {description: 'All fields are required'}, '403': {description: 'Access denied. You do not have permission to update this workout.'}, '404': {description: 'Workout not found'}, '422': {description: 'Invalid input values'}}
        delete:
            summary: 'Delete a specific workout'
            tags: [Workouts]
            parameters: [{name: workout_id, in: path, required: true, description: 'ID of the workout', schema: {type: integer}}]
            responses: {'200': {description: 'Workout deleted successfully'}, '403': {description: 'Access denied. You do not have permission to delete this workout.'}, '404': {description: 'Workout not found'}}
...